Furniture encapsulates Routes and Views
See: https://github.com/zinc-collective/convene/issues/186

One thing that was concerning me was the way that Furniture code
definitions was starting to spread throughout the codebase, particularly
how if someone wanted to define a piece of furniture, they would have to
define stuff in `app/views` as well as `app/lib/furniture` and
`config/routes.rb`

This consolidates all those entrance points, so that Furniture can be
defined in a relatively localized place.
