+++ convene-web/app/lib/furniture/base_controller.rb
@@ -1,4 +1,5 @@
 module Furniture
   class BaseController < ::ApplicationController
+    prepend_view_path 'app/lib'
   end
 end
