+++ convene-web/app/lib/furniture/breakout_tables_by_jitsi.rb
@@ -1,19 +1,9 @@
 module Furniture
-  class BaseController < ::ApplicationController
-  end
-
-  class BreakoutTables
-    attr_accessor :placement
-
-    delegate :settings, to: :placement
-
-    def initialize(placement)
-      self.placement = placement
+  class BreakoutTablesByJitsi
+    def self.append_routes(router)
+      router.resources :breakout_tables, only: [:show], controller: 'breakout_tables_by_jitsi/'
     end
-
-    def in_room_template
-      'furniture/breakout_tables/in_room'
-    end
+    include Placeable
 
     def find_table(table_name)
       tables.find { |table| table.name == table_name }
@@ -22,21 +12,6 @@ module Furniture
     def tables
       settings['names'].lazy.map do |name|
         Table.new(name: name, placement: placement)
-      end
-    end
-
-    class Controller < Furniture::BaseController
-      def page_title
-        "#{room.space.name}, #{room.name}, #{table.name}"
-      end
-
-      helper_method def table
-        @table ||= room.furniture_placements.find_by!(name: :breakout_tables)
-          &.furniture&.find_table(params[:id])
-      end
-
-      helper_method def room
-        @room ||= current_space.rooms.friendly.find(params[:room_id])
       end
     end
