+++ convene-web/app/lib/furniture/breakout_tables_by_jitsi/_in_room.html.erb
