+++ convene-web/app/lib/furniture/breakout_tables_by_jitsi/show.html.erb
