+++ convene-web/app/lib/furniture/videobridge_by_jitsi.rb
@@ -1,0 +1,6 @@
+module Furniture
+  # Provides an iFramed Jitsi Meet to a {Room}.
+  class VideobridgeByJitsi
+    include Placeable
+  end
+end
