--- convene-web/app/lib/furniture/videobridge_jitsi.rb
@@ -1,9 +1,0 @@
-module Furniture
-  class VideobridgeJitsi
-    def initialize(placement)
-      def in_room_template
-        "furniture/videobridge_jitsi/in_room"
-      end
-    end
-  end
-end
