+++ convene-web/config/application.rb
@@ -1,18 +1,18 @@
 require_relative 'boot'
 
-require "rails"
+require 'rails'
 # Pick the frameworks you want:
-require "active_model/railtie"
-require "active_job/railtie"
-require "active_record/railtie"
-require "active_storage/engine"
-require "action_controller/railtie"
-require "action_mailer/railtie"
-require "action_mailbox/engine"
-require "action_text/engine"
-require "action_view/railtie"
-require "action_cable/engine"
-require "sprockets/railtie"
+require 'active_model/railtie'
+require 'active_job/railtie'
+require 'active_record/railtie'
+require 'active_storage/engine'
+require 'action_controller/railtie'
+require 'action_mailer/railtie'
+require 'action_mailbox/engine'
+require 'action_text/engine'
+require 'action_view/railtie'
+require 'action_cable/engine'
+require 'sprockets/railtie'
 # require "rails/test_unit/railtie"
 
 # Require the gems listed in Gemfile, including any gems
@@ -32,6 +32,8 @@ module ConveneWeb
     # Don't generate system test files.
     config.generators.system_tests = nil
 
+    # Allows us to place our views for pieces of furniture next to the furniture
+    config.paths['app/views'].unshift(Rails.root.join('app', 'lib', 'furniture'))
     config.session_store :cookie_store, expire_after: 7.days
   end
 end
