+++ convene-web/spec/models/furniture_placement_spec.rb
@@ -20,7 +20,7 @@ RSpec.describe FurniturePlacement do
       furniture_placement = FurniturePlacement
                             .new(name: :breakout_tables, settings: { names: %w[a b] })
 
-      expect(furniture_placement.furniture).to be_a(Furniture::BreakoutTables)
+      expect(furniture_placement.furniture).to be_a(Furniture::BreakoutTablesByJitsi)
       expect(furniture_placement.furniture.placement).to eql(furniture_placement)
       expect(furniture_placement.furniture.tables.to_a[0].name).to eql('a')
       expect(furniture_placement.furniture.tables.to_a[1].name).to eql('b')
